# fix performance degradation on handling conflict fields