--- a/src/graphql/validation/rules/overlapping_fields_can_be_merged.py
+++ b/src/graphql/validation/rules/overlapping_fields_can_be_merged.py
@@ -11,6 +11,7 @@ from ...language import (
     ObjectFieldNode,
     ObjectValueNode,
     SelectionSetNode,
+    ValueNode,
     print_ast,
 )
 from ...type import (
@@ -557,7 +558,7 @@ def find_conflict(
             )
 
         # Two field calls must have the same arguments.
-        if stringify_arguments(node1) != stringify_arguments(node2):
+        if not same_arguments(node1, node2):
             return (response_name, "they have differing arguments"), [node1], [node2]
 
     directives1 = node1.directives
@@ -596,16 +597,34 @@ def find_conflict(
 
     return None  # no conflict
 
+def same_arguments(node1: Union[FieldNode, DirectiveNode], node2: Union[FieldNode, DirectiveNode]) -> bool:
+    args1 = node1.arguments
+    args2 = node2.arguments
 
-def stringify_arguments(field_node: Union[FieldNode, DirectiveNode]) -> str:
-    input_object_with_args = ObjectValueNode(
-        fields=tuple(
-            ObjectFieldNode(name=arg_node.name, value=arg_node.value)
-            for arg_node in field_node.arguments
-        )
-    )
-    return print_ast(sort_value_node(input_object_with_args))
+    if args1 is None or len(args1) == 0:
+        return args2 is None or len(args2) == 0
 
+    if args2 is None or len(args2) == 0:
+        return False
+
+    if len(args1) != len(args2):
+        return False
+
+    values2 = {arg.name.value: arg.value for arg in args2}
+
+    results = []
+    for arg1 in args1:
+        value1 = arg1.value
+        value2 = values2.get(arg1.name.value)
+        if value2 is None:
+            results.append(False)
+        else:
+            results.append(stringify_value(value1) == stringify_value(value2))
+
+    return all(results)
+
+def stringify_value(value: ValueNode) -> str:
+    return print_ast(sort_value_node(value))
 
 def get_stream_directive(
     directives: Sequence[DirectiveNode],
@@ -626,7 +645,7 @@ def same_streams(
         return True
     if stream1 and stream2:
         # check if both fields have equivalent streams
-        return stringify_arguments(stream1) == stringify_arguments(stream2)
+        return same_arguments(stream1, stream2)
     # fields have a mix of stream and no stream
     return False