When a crafted query is received, it significantly exhausts computing resources like the CPU, which negatively impacts response time.
For example, adding the following benchmark takes about 7 ~ 8 seconds in my environment (M1 Mac / 32GB).

```python
from graphql import build_schema, parse, validate
from graphql.utilities import get_introspection_query

from ..fixtures import big_schema_sdl  # noqa: F401

def test_validate_conflict_fields_query(benchmark, big_schema_sdl):  # noqa: F811
    schema = build_schema(big_schema_sdl, assume_valid=True)
    fields = "__typename " * 1000
    query = parse('query {{ {fields} }}'.format(fields=fields))
    result = benchmark(lambda: validate(schema, query))
    assert result == []
```

```shell
$ time poetry run pytest tests/benchmarks/test_validate_gql_conflict_fields.py
Executed in    8.23 secs    fish           external
   usr time    7.60 secs    0.06 millis    7.60 secs
   sys time    0.22 secs    1.01 millis    0.21 secs
```

This is a similar problem in [graphql-js and has been fixed](https://github.com/graphql/graphql-js/pull/3958).  
This PR has been ported to make the same changes.

This change will greatly improve performance:

```shell
$ time poetry run pytest tests/benchmarks/test_validate_gql_conflict_fields.py
Executed in    1.93 secs    fish           external
   usr time    1.47 secs    0.07 millis    1.47 secs
   sys time    0.18 secs    1.22 millis    0.18 secs
```