@Cito 2024-02-14
A release with this fix will be available soon.