@datur 2024-06-04
Hey @Cito 👋  do you know when you might be putting up a release that includes this fix? If there's anything I can do to help it along let me know