@Cito 2024-06-04
@datur It's already contained in v3.3.0a5. It's still labelled as an alpha release because the corresponding GraphQL.js version 17 is also still labelled as alpha.